cros_ec/mkbp: Check interrupt line before polling keyboard state
This patch fills the cros_ec_interrupt_pending() function with life that
has apparently always been there but never used. It can now be wired up
to a GPIO which is then checked before polling the EC for keyboard
input. This will make the AP only poll the EC when there's actually
keyboard input to receive, which significantly reduces the EC's
processing overhead and should avoid problems with starving other tasks.

Hook up this feature as an example on Veyron boards (however, due to the
state of the project it will not be ported back into the stable branch).

BRANCH=none
BUG=chrome-os-partner:39514
TEST=Booted Jerry, confirmed that it could still detect keypresses.
Confirmed that EC log does not show a huge amount of MKBP polls.

Change-Id: I02ac3fa5e03ed36212d99e339fad33b4b3e0b63a
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/267354
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>