mkbp: Repair broken keyboard FIFO handling
Commit 2594734 (cros_ec/mkbp: Check interrupt line before polling
keyboard state) introduced a subtle flaw in the mkbp keyboard driver,
which allows a bookkeeping routine to abort too early and leave the key
input FIFO in an inconsistent state for getchar() (which will then
return a bunch of uninitialized memory bytes as "key presses"). This
only happens after the first key press is parsed correctly, so it fell
through cursory testing.

This patch fixes the issue by moving the new EC interrupt early abort
check after the point where the FIFO gets reinitialized to a consistent
state. It also changes havechar() to check for fifo_offset < fifo_size
instead of just returning fifo_size. Even though this should only be
possible if fifo_size is 0 with the fixed code, it's still a more
accurate representation of "FIFO is empty" (and would've also prevented
this bug). (Also generally use fifo_offset < fifo_size instead of
fifo_offset != fifo_size to be safer against unintended overflows like
this in the future.)

BRANCH=none
BUG=chrome-os-partner:39514
TEST=Booted Jerry, made sure that it would detect keys correctly and
one-at-a-time, even for multiple presses.

Change-Id: I54f1ac903ab213df0b0ecbd702ed32ff4cedc90a
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/269804
Reviewed-by: David Hendricks <dhendrix@chromium.org>