fastboot: Correct fastboot getvar response
fastboot host expects device to send back OKAY for all the variables
even if they are not identified by the device. It expects the device
to just put in "unknown command or N/A" string for certain variables
that are not defined by the board.

This patch corrects the response type for variables that are not
defined to FB_OKAY instead of FB_FAIL and updates the response message
to "unknown variable". Additionally, it corrects the response of
product name to return value of hardware_name() function.

BUG=None
BRANCH=None
TEST=Compiles successfully and fastboot update works fine.

Change-Id: I32fe6b1862e0827fce4d4ea68d471cb5cb31b5c0
Signed-off-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/271379
Tested-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>
Trybot-Ready: Furquan Shaikh <furquan@chromium.org>