vboot: Localize flash_nvram_init() to vboot code
This makes VbExNvStorageRead() and VbExNvStorageWrite() call
flash_nvram_init() which will initialize itself, if needed, or
simply return if it's already been initialized.

The idea is to remove the dependency on having each board with
flash-backed NVRAM (which will be most/all of them in the future)
from having to call flash_nvram_init() explicitly.

BUG=chrome-os-partner:41298
BRANCH=none
TEST=veyron_mickey can now switch in and out of dev mode.

Change-Id: I8014d7d102b02a05735874fc79484f74147e563c
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/277440
Reviewed-by: Julius Werner <jwerner@chromium.org>
Tested-by: Jiazi Yang <Tomato_Yang@asus.com>
Reviewed-on: https://chromium-review.googlesource.com/285858