fit: Properly handle DT device_type = "memory" interface
When updating the /memory node in the device tree, we currently just
look it up  by node name. However, this is not actually the deciding
factor specified in Linux' Documentation/devicetree/booting-without-
of.txt, and in fact some sources recommend naming the node
/memory@<address>, which won't get matched by depthcharge.

The deciding factor for the kernel is the "device_type" property which
must be set to "memory". It even allows multiple of those nodes and
would merge them together. Depthcharge is currently not even adding the
device_type property, so if it doesn't find an existing /memory node
with it then the one it creates will be ignored by the kernel.

This patch changes the update_memory() function to explicitly find and
delete all memory nodes by device_type first, and then create a single,
fully filled out node that will be found by the kernel.

BRANCH=None
BUG=chrome-os-partner:41728
TEST=Booted Oak, confirmed that /proc/device-tree contained only one
memory node with the correct range and that /proc/iomem correctly
excludes the coreboot table carveout from system memory.

Change-Id: Id99799e71b4aaaaecef202bd7dc72834401cc011
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/281442
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/282142
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>
Trybot-Ready: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>