config: Bump default KERNEL_SIZE to 32M
This patch increases the buffer size for the kernel image to 32M. While
the original 16M should usually be enough for normal Chrome OS use, some
developer features (e.g. fastboot, netboot, especially if you add a
ramdisk) might want a little more. There's no real penalty to increasing
this (it's not part of the BSS), so let's bump it up to 32M (which is
unfortunately the best we can do without hacking up a bazillion
defconfigs).

Also update related Kconfig descriptions to be a little closer to
current practice.

BRANCH=None
BUG=None
TEST=None

Change-Id: Ibbb35a43007108289e952c27f165ee65da78e783
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/281806
Reviewed-by: Aaron Durbin <adurbin@chromium.org>