fastboot: Add support for off-mode-charge
When off-mode-charge argument is set to 1, by default, device is
supposed to only charge the battery in off-state. However, if the
argument is set to 0, then device should start booting as soon as AC
is plugged in off-state.

In response to off-mode-charge fastboot command, set boot_on_ac_detect
flag in EC as well in NVRAM storage(to ensure that the state is not
lost even in battery drain conditions).

CQ-DEPEND=CL:289900
BUG=chrome-os-partner:41680
BRANCH=None
TEST=Compiles successfully and AP boots up on plugging AC power in off
state with flag set.

Change-Id: Ib8ce836d9888dd4c0448c168339b66c3b4cc8260
Signed-off-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/290026
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Trybot-Ready: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>