net: Add 1 mS delay before polling for USB net device
This is a temporary workaround until the real root cause
is found.

It is found that in strago/cyan devices, sometimes USB ethernet
device is not enumerated, as the SET_ADDRESS command
fails during enumeration. While the LeCroy USB captures
show no activity on the bus, the debug messages in libpayload
indicate failures around set_address command. Adding a small
delay (1 mS), before polling for the USB ethernet device,
gets the enumeration to go through successfully. This delay
is needed only so long as the USB ethernet device enumeration
is not done.

BUG=chrome-os-partner:43941
BRANCH=firmware-strago-7287.B
TEST=Build netboot image and verify netboot works fine.

Change-Id: Ibe6a6ef7978bd22b83aaebfb8464c0007ee61a94
Signed-off-by: Rajmohan Mani <rajmohan.mani@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/293579
Reviewed-by: Aaron Durbin <adurbin@chromium.org>