veyron_rialto: implement firmware LED behavior
Add the capability to show the various modes for depthcharge through the limited
set of leds.

During a normal boot all leds are on (to show the user that the device is
working), but they'll change when the userspace software starts doing its
thing. If the device gets stuck in firmware, the error led (warning icon)
will be on and some leds will be off according to the following pattern:

* only error: recovery
* error + pink wireless: bad usb image
* error + orange wireless: remove USB flash drive
* error + sync: dev mode
* error + sync + white wireless: switching to/from dev mode
* all leds: either booting normally or just confirmed to normal transition

Note that the orange wireless icon(alone) will be turned on for a split second
before coreboot gets a chance to load, this is a hardware issue unfortunatelly.

BUG=chrome-os-partner:44274
TEST=Go through the recovery/dev mode flow, note the various led patterns
BRANCH=master

Change-Id: I516d59651097f10319577af5c3a1b8d990a79877
Signed-off-by: Alexandru M Stan <amstan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/294668
Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/304303
Commit-Queue: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>