spi: Make error checking more generic
The error checking that was originally implemented is too specific to
Spansion SPI flash chips (and similar ones, if they exist). This
adds logic to read the chip's manufacturer and device ID via the
RDID opcode (0x9f) so that we can check errors appropriately for
more chips.

Currently this only handles Spansion (as earlier) and does nothing
for other chips.

BUG=chrome-os-partner:44433
BRANCH=firmware-smaug
TEST=needs testing

Change-Id: Ic4408501b3483ffa2813857bd278f53274a2b0fe
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/295630
Reviewed-by: Furquan Shaikh <furquan@chromium.org>