gpio: skylake: Add support for 'fast' GPIO setting
In order to do bit-bang operations as fast as possible it is necessary
to avoid reads in the GPIO set path.  This requires maintaining the
current state of the DW0 register instead of reading it every time.

Additionally the GPIO setup routines have been modified to actually
configure the GPIO as an input/output if that specific type of GPIO
is requested.  The original GPIO state is then restored when exiting
depthcharge.

The basic "new_skylake_gpio" without input/output specified does not
do any modification to the GPIO configuration or attempt to restore
the existing state on exit.

BUG=chrome-os-partner:42283
BRANCH=none
TEST=build and boot on glados

Change-Id: I90261307c12161c3af23b0b407475d44cc7fe447
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/301382
Reviewed-by: Aaron Durbin <adurbin@chromium.org>