glados: Add support for audio beep
Add support for beeping by using GPIO based PDM output to the
SSM4567 audio amplifier.  Only the left speaker is used because
the PDM output stream is not very clean and using both speakers
can cause the audio to not be synchronous on both speakers.

The sample rate is tuned to make the tone sound appropriate given
the slow rate of the PDM clock.  The clock start value is tuned
to reduce the audio startup pop and crackling as much as possible.

The GPIOs that are used are GPP_F0 and GPP_F2 and they are configured
to be outputs but then will be returned to input when depthcharge
exits.

BUG=chrome-os-partner:42283
BRANCH=none
TEST=test audio beep on glados

Change-Id: I6d1385de9300ad1e455a9a76effc76ac20142b32
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/301385
Reviewed-by: Aaron Durbin <adurbin@chromium.org>