vboot: support locales
This change adds locale parameter to vboot_draw_screen so that it can
draw locale dependent screens. When initializing, it reads 'locales'
from cbfs, in which supported locales are listed.

If no locales are found or no 'locales' file is found, it falls back to
the debug strings.

BUG=chromium:502066
BRANCH=tot
TEST=Tested on Samus
CQ-DEPEND=CL:304375

Change-Id: Idd54064829df564c0fa01e24e20cb357798ff005
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/304382