fastboot: Define fastboot callback handler
Instead of defining stubs for each possible callback handler, define a
callback table that can be used by boards to define whatever callbacks
are implemented. This helps in following ways:
1. Avoid adding of stubs for every new callback
2. Allows re-use of common functions like calls to ec. This avoids
putting in any hard-coded callbacks to ec for performing tasks like
battery cut-off, boot-on-ac-detect and others.

Downside of this approach is that before making a call to the callback
handler, we need to ensure that the callback actually exists.

This change is being made to make the fastboot component ready for
merging all changes from smaug branch in a clean way that can be
useful for other boards if required.

Also update callback tables for smaug, rush_ryu and foster.

BUG=None
BRANCH=None
TEST=Compiles successfully.

Change-Id: Iae852f9b01d0ff9ffbb2254fbee48901c58536c7
Signed-off-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/305231
Commit-Ready: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>