flash driver: provide setup for selecting cbfses from fmap
The offset needn't be added at all times (in fact, just for the first
two accesses), so we look for out of bounds conditions to check when
it's required to add the offset and when it needn't be.

This approach works because the CBFS images' offsets are larger than
their own size. This restriction (and this imprecise approach) will be
lifted in follow-up commits.

BUG=chromium:445938
BRANCH=tot
TEST=none

Change-Id: I3f60a3a66ed8d6654be5bed1ca1f4b4b2cd5b8d7
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/306600
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>