Support fetching EC firmware from CBFS-formatted FW_MAIN_*
With MULTIPLE_CBFS_INSTANCES enabled, EC firmware and hashes
come from CBFS instead of its separate fmap region and an indexes
FW_MAIN_*, respectively.

This is currently implemented as a fallback option when the normal
"indexed" approach fails. Doesn't hurt to add another try then, and due
to details in the CBFS format, it's guaranteed that this fallback is
taken with MULTIPLE_CBFS_INSTANCES.

Once that configuration is in operation, the old code path can be taken
out.

BUG=chromium:445938
BRANCH=tot
TEST=none

Change-Id: I5ac553de3f5dcd0bf59cc1e91aa2e7eed6c23e6e
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/306602
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>