vboot: ec: Check for EC limit power request in VbExEcVbootDone()
Add VbExEcVbootDone(), which checks the limit power flag that gets set by
the EC to request that the AP limit its power use. If this flag remains
set for 3 seconds, request a system shutdown since we're likely to
brownout if we continue to boot.

BUG=chromium:537269
TEST=Manual on Glados. Set CHG_MW thresh to 20000, BAT_PCT to 50. Verify
that LIMIT_POWER parameter is set until Zinger negotiates to 20V. Also
verify that we do not proceed with boot when Donette is plugged.
BRANCH=None

Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>
Change-Id: I524e7eb5674f2061791cca920c8e1b4c35805d3c
Reviewed-on: https://chromium-review.googlesource.com/307885
Commit-Ready: Shawn N <shawnn@chromium.org>
Tested-by: Shawn N <shawnn@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>