storage: ahci: implement drive firmware update mechanism
Support device (e.g., SSD) firmware upgrades via DOWNLOAD MICROCODE ATA
command (mode 3 -- chunked upload -- only). Updates can be provided by
the board file, along with some metadata about when to apply the update.
We match against the model and firmware revision strings found in the
IDENTIFY payload.

Because the buggy Kingston firmware update process requires the
drive to be power cycled for the update to take effect, we provide a
platform callback along with the update payload, for implementing the
SSD power cycle.

Use this for samus only.

BUG=chromium:467131
TEST=test Kingston SSD update from 1.8 to 2.3 firmware on samus

Change-Id: I83a76c91fc11e47347bd89bc53d2465e45c8664c
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/308077
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>