vboot: fix text heights getting smaller
get_text_width calls get_image_size repeatedly for each letter in a text
to calculate the total width of the text being printed on screen.
Since it reuses the height returned from get_image_size, the height can
get smaller and smaller because of integer rounding. This patch fixes the
issue by not carrying over the returned height to the next iteration.

BUG=chromium:502066
BRANCH=tot
TEST=Tested on Glados

Change-Id: I129123997c1f539d9248bd25fdb91120dcf02395
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/309771
Reviewed-by: Randall Spangler <rspangler@chromium.org>