graphics: Do not enable backlight in graphics_init
Instead of enabling backlight in graphics init, let the caller decide
when to enable/disable the backlight. Currently, the flows in vboot
follow different paths for displaying screens depending upon whether a
screen is present in GBB or not. Adding a backlight enable in graphics
init breaks the sequence for vboot.

BUG=chrome-os-partner:43706
BRANCH=None
TEST=Compiles successfully

Change-Id: I550b6f481e1b1659d475ec93a952e4f5bad33c00
Signed-off-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/312747
Commit-Ready: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>