glados: Workaround for silego unable to see EC reset
On skylake boards the silego chip is unable to see EC resets and
clear the latch on EC_IN_RW state.  This means that FAFT is unable
to switch to developer mode because VbExTrustEC() will always fail.

This only matters for FAFT because entering recovery mode with the
keyboard still works properly and this VbExTrustEC() callback is
only ever used in the developer mode transition screen.

As such, use the existing FAFT override GBB flag that is used for
similar workarounds in the firmware to enable FAFT testing.

This is a bit messy, using an overridden flag to check at runtime,
because GBB flags are not available in the board_setup() function
as they have not been read yet and due to the structure of RO/RW
depthcharge it is not convenient to change that behavior.

BUG=chrome-os-partner:47648
BRANCH=none
TEST=manual testing (in additon to FAFT)
1) set GBB flag 0x300
2) enter recovery mode with: dut-control power_state:rec
3) press Ctrl+D on keyboard to transition to developer mode
4) also verify that when the GBB flag is not set that if the
EC is in RW and steps 2-3 are executed that it does not allow
entry into developer mode
5) also verify that keyboard controlled recovery mode does clear
the EC_IN_RW flag properly and does not need GBB flag to be set.

Change-Id: I30e870e16e63f34a7f63c9e977134a6d5d27cbe1
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/319245
Reviewed-by: Aaron Durbin <adurbin@chromium.org>