Kunimitsu/Lars: Workaround for silego unable to see EC reset
On skylake boards the silego chip is unable to see EC resets and
clear the latch on EC_IN_RW state.  This means that FAFT is unable
to switch to developer mode because VbExTrustEC() will always fail.

This only matters for FAFT because entering recovery mode with the
keyboard still works properly and this VbExTrustEC() callback is
only ever used in the developer mode transition screen.

As such, use the existing FAFT override GBB flag that is used for
similar workarounds in the firmware to enable FAFT testing.

This is a bit messy, using an overridden flag to check at runtime,
because GBB flags are not available in the board_setup() function
as they have not been read yet and due to the structure of RO/RW
depthcharge it is not convenient to change that behavior.

This is a port for Kunimitsu and lars boards following CL for
https://chromium-review.googlesource.com/#/c/319245/

BUG=chrome-os-partner:47648
BRANCH=none
TEST=manual testing (in additon to FAFT)
1) set GBB flag 0x300
2) enter recovery mode with: dut-control power_state:rec
3) press Ctrl+D on keyboard to transition to developer mode
4) also verify that when the GBB flag is not set that if the
EC is in RW and steps 2-3 are executed that it does not allow
entry into developer mode
5) also verify that keyboard controlled recovery mode does clear
the EC_IN_RW flag properly and does not need GBB flag to be set.

Change-Id: Icd120cbff4d19b51f193c5ed9c40ec3098820f5e
Signed-off-by: Naveen Krishna Chatradhi <naveenkrishna.ch@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/319208
Commit-Ready: Divya Jyothi <divya.jyothi@intel.com>
Tested-by: Divagar Mohandass <divagar.mohandass@intel.com>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>