vboot: try rw-cbfs first then fall back to ro-cbfs
This CL makes depthcharge search bitmap images in rw-cbfs first. If it's
not found it falls back to ro-cbfs. In this way, we can update screens
in the field and use spaces in rw-cbfs only when needed.

This feature is gated on new Kconfig flag MEDIA_RESOURCES_IN_RW, which
can be enabled by the depthcharge ebuild.

BUG=b:112640892
BRANCH=eve-campfire
TEST=Compile chromeos-bmpblk with USE=bmpblk_rw.
     Compile firmware image and update AP-RW.
     Check that screen resources are properly located.

Change-Id: Idd74f6dd7702eba9f0a0229a96822c6e9d937a8c
Reviewed-on: https://chromium-review.googlesource.com/1177288
Commit-Queue: Joel Kitching <kitching@chromium.org>
Tested-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>