rialto: Increase dev mode delay via CUSTOM_MUSIC
Define the CUSTOM_MUSIC variable, this allows us to provide VbExGetMusicPtr
and VbExMaxMusicSize which reference a "song" memory block.

A new "song" was crafted by adding lots of silent places.

The beeps(which can't be heard on rialto anyway) together had to be longer
than 1 second, otherwise some code would judge my song and deem it
unworthy of playing.

Since VbDevMusicNote and VbDevMusic typedefs were in a vboot_audio_private.h,
I had to hardcode the song in a simple array without the niceties of a struct.
The crc was found by compiling once and copying it from the checksum
mismatch error.

TEST=Compile and flash, note dev mode is much longer now
BUG=chrome-os-partner:48657
BRANCH=master

Change-Id: If823f0d2c083a7f3102115f5ef227f16c49a5895
Signed-off-by: Alexandru M Stan <amstan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/321722
Commit-Ready: Alexandru Stan <amstan@chromium.org>
Tested-by: Alexandru Stan <amstan@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/324815