skylake boards: prepare for separate verstage
In order for the skylake boards to have a separate verstage
and run romstage from the RW slots one has to ensure the FSP
and romstage are added to CBFS for XIP (execute-in-place).
This requires one to re-add FSP and romstage so that the
two files are relocated. Additionally, remove verstage
from the RW slots since verstage would never be run from the
RW slots.

BUG=chrome-os-partner:44827
BRANCH=glados
TEST=Built and booted chell using separate verstage. Also, suspended
     and resumed.

Change-Id: I145d4c85755f0b885bcf7e59b7284c2a60993303
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/323421
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>