mmc: sdhci: make sure get the coherent/uncached memory when do dma transfer
in ARM platform we need to worry about cache coherency when accessing
DMAed data. This patch enhances SDHCI driver to ensure validity of
data in the memory shared with the controller.

bouncebuf.c is edited to make it compile properly on x86 platforms, as
it is now included by the make.

BUG=None
BRANCH=None
TEST=None

Change-Id: I9e33893ca5d8e153250f2249ced0fbe5cecc8223
Signed-off-by: Lin Huang <hl@rock-chips.com>
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/330283
Reviewed-by: Julius Werner <jwerner@chromium.org>