commonparams: Fix issue with gbb_init and add gbb_get_flags
Add a new function for gbb_get_flags() to return the GBB flags
for the image.

This (and other) GBB utility function are able to be called before
common_params_init() as they call gbb_init() directly, but if
common_params_init() is called after gbb_init() then the GBB data
set in cparams will be silently cleared and not set again because
gbb_init() does not run again.

Save the GBB size and data pointer before the memset and restore
after to preserve the GBB data and allow these functions to be
safely called if common_params_init() has not been run yet.  This
is useful since common_params_init() is not run until after all
INIT_FUNC have been run so those functions were not able to safely
access the GBB without potentially breaking vboot behavior.

BUG=chrome-os-partner:50864
BRANCH=glados
TEST=test lid disable GBB flag can be read in netboot image which
does not call common_params_init(), and that it can be read successfully
both in an init function and in the vboot loop for a normal image.

Change-Id: I228d2c4ee361319f431f4e63d3b7ddcf3cc725cd
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/331306
Reviewed-by: Aaron Durbin <adurbin@chromium.org>