cros_ec: Add init function to disable LID shutdown event
If the GBB flag is set to disable lid shutdown we still can
receive an SMI and shut down the system.

We can however disable this SMI event for lid close once we
can read the GBB flags which is here in depthcharge.  This still
has a window where the SMI event is enabled in coreboot but not
yet disabled in depthcharge, but by the time the firmware screen
is displayed it should be safe to shut the lid.

The SMI event mask is only applicable to X86 systems so this is
set to only be included for CONFIG_ARCH_X86.

BUG=chrome-os-partner:50864
BRANCH=glados
TEST=test that lid shutdown SMI events are disabled in depthcharge
if the corresponding GBB flag is set.

Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/331308
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
(cherry picked from commit 6ae79a4a1723dafbc5fc88e6271316e158b17e66)
Change-Id: I1d6483988ffcef762feba6049d3b7da93180e549
Reviewed-on: https://chromium-review.googlesource.com/331294