board/lars: Remove the workaround for silego unable to see EC reset
With the new Silego fix on the DVT boards. We can remove the software
workaround for LARS boards.

BUG=chrome-os-partner:49847
BRANCH=glados
TEST=manual testing (in additon to FAFT)
1) set GBB flag 0x300
2) enter recovery mode with: dut-control power_state:rec
3) press Ctrl+D on keyboard to transition to developer mode
4) also verify that when the GBB flag is not set that if the
EC is in RW and steps 2-3 are executed that it does not allow
entry into developer mode
5) also verify that keyboard controlled recovery mode does clear
the EC_IN_RW flag properly and does not need GBB flag to be set.

Change-Id: I14dbb7c0731853065dce724a6c3294626c2dd71d
Signed-off-by: Naveen Krishna Chatradhi <naveenkrishna.ch@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/331366
Commit-Ready: Naveenkrishna Ch <naveenkrishna.ch@intel.com>
Tested-by: Naveenkrishna Ch <naveenkrishna.ch@intel.com>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>