vboot: Draw bad USB icon in the center
When the device doesn't support SD card recovery, BadSD.bmp is created as
an empty image. When Depthcharge renders a screen, it shows BadSD.bmp and
BadUSB.bmp side by side, resulting in USB stick image shifted to the right.

This change fixes the issue by showing BadDevices.bmp, which can be either
BadDevices.png or BadUSB.png, depending on the board configuration. This is
consistent with how we render InsertDevices.bmp.

BUG=chromium:501060
BRANCH=tot
TEST=Tested on Jerry in 50 languages with and without SD card support
CQ-DEPEND=CL:337599

Change-Id: I4f70701e8d35827f338250c7a9521313461c1c8b
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/337609
Reviewed-by: Aaron Durbin <adurbin@chromium.org>