Log kernel command line in firmware console
Whenever I get a bug report of a machine hanging after "Exiting
depthcharge..." I tell them it's a kernel problem and they forgot to
enable earlyprintk, and 99% of the time I'm right. Still, in order to
make that argument easier and to help debug other issues where the
kernel's dump of its own command line may not be visible, let's just
print it out from firmware as well.

BRANCH=None
BUG=None
TEST=None

Change-Id: I449bac72d5204ff651dd0fdc47debca53c5b29a8
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/334493
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>