anx7688: Add support for ANX7688
ANX7688 is a PD + HDMI->DP converter. It contains a firmware
that we update from the AP-FW, at boot time, which is the only
reason to have a driver for it in depthcharge.

We reuse the PD software sync mechanism, with some caveats as
the chip does not have proper RO/RW sections.

BRANCH=none
BUG=chrome-os-partner:52434
TEST=Boot elm-rev0

Change-Id: I4e16dfdfea236a0de5c9f6881e39513457402971
Signed-off-by: Nicolas Boichat <drinkcat@google.com>
Reviewed-on: https://chromium-review.googlesource.com/339497
Commit-Ready: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>