checkpatch: Ignore LINE_SPACING
The newest checkpatch version now flags code such as

	int i;
	for (i = 0; ...

because there is no blank line between a declaration and the next
statement. We have always allowed variable declarations inline with the
code in depthcharge, so this doesn't make much sense for us. Add it to
the ignore list.

BRANCH=None
BUG=None
TEST=None

Change-Id: I982edc503dfd4870a20bf59d5aadacb0aa4f5df7
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/341474
Reviewed-by: Aaron Durbin <adurbin@chromium.org>