config: Ensure sane minimum heap sizes across boards
While some boards use the standard 32MB heap size, some explicitly limit
it to lower values. (I'm not quite sure if all of those are really
deliberate or just the result of copy&paste.) Keeping the heap small has
a little boot time benefit (for zeroing out less memory), but overdoing
it runs the risk of OOM issues. The advent of CBFS-based imagery seems
to have increased our memory requirements by a bit (with the
memory-mapped vbgfx.bin alone taking over 1MB), so that a 6MB heap on a
board that already needs 4MB to mirror the SPI flash is dangerously
close to exhaustion. This patch increases all those heaps to 8MB which
ought to be enough for anybody.

BRANCH=None
BUG=chrome-os-partner:52554
TEST=I can keep cycling through firmware languages for several minutes
without any bugs or crashes, despite a small memory leak.

Change-Id: Iac42867d4ab14a61247a37449c2cd59a52d84bdd
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/346216
Tested-by: Yidi Lin <yidi.lin@mediatek.com>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>