reef: increase BIOS region size
The new descriptors utilize a 512KiB device expansion region for the
CSE's use. This allows the BIOS region to increase as the expansion
region decreased. Adjust the fmap as well as the macro marking the
end of the BIOS region. The CSE_SIGN region needed to be moved to
reflect the new location of the logical boot partition 2 since
the BIOS region increased. The GBB also had to be decreased to
accommodate the new LBP2 location.

BUG=chrome-os-partner:53689
BRANCH=None
TEST=None

Change-Id: Ib40335b9c0adfa824cec30768a87e12007735e59
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/347482
Reviewed-by: Furquan Shaikh <furquan@chromium.org>