vboot: ec: Store vboot hash for EC image in a separate file
CBFS attributes are mostly useful for uses within CBFS itself, such as
the transparent compression. Using the CBFS hash attribute for vboot
causes multiple problems: it forces us to use hashes that CBFS can
understand and generate on its own (whereas for some of our newer PD
chips we'd like to feed vboot a "hash" that's not really a hash at all,
but just a version number), and it prevents us from compressing the EC
images in CBFS (since the CBFS hash attribute was originally designed
for verification and thus always hashes the compressed data). Things
will become easier if we just split these two distinct things apart.

This patch instead just assumes that for every EC image file X, there is
also a file X.hash which contains the vboot hash for it (e.g.
"ecrw.hash"). When vboot asks us for the hash we look up this hash file,
and when it asks us for the full image we look up the normal image file.
The hashes will be precomputed by the build system in a way that allows
customization per chip type.

CQ-DEPEND=CL:348054
BRANCH=None
BUG=chrome-os-partner:53780
TEST=Forced software sync on Oak in multiple scenarios.

Change-Id: I3bfcc8eaa54c4559fe7d4c28dd74f030ebc8455c
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/348061
Commit-Ready: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>