gru: consolidate gru based boards in common board directory
Kevin and Gru are very similar from the AP firmware point of view, the
only difference is the fact that on Gru it is required to bring up
USB2 interfaces for booting.

A new configuration option is being added to allow to fine tune this
option at the compile time and the rest of board configurations is
being synchronized.

BRANCH=none
BUG=none
TEST=built firmware for both Keving and Gru, booted into the Linux
     command line and successfully ran stressapptest for an hour.

Change-Id: Ifb57a0c3d50a8ca9a7a5acad0c8311e2f2c93d03
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/354250
Reviewed-by: Julius Werner <jwerner@chromium.org>