drivers/flash: Support mmaped flash with non-zero base offset
In some boards e.g. reef, the entire flash is not memory mapped. Instead
only a part of the flash (offset 0x1000 to offset x which covers the
BIOS region) is mapped in the address space. In order to handle such
special cases, a new API is required to allow boards to define the
offset on flash of the base address that is memory mapped. This base
offset needs to be taken care of while accessing any absolute address
within the mmaped range. For older boards, this base_offset defaults to
0.

BUG=chrome-os-partner:54563
BRANCH=None
TEST=Compiles successfully and boots to OS on reef.

Change-Id: I8cc54168235ceeaf561c7e77c0b661ed053bcd4b
Signed-off-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/354380
Commit-Ready: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>