Don't generate .payload files.
When the switch to CBFS for all the RW slots the .payload files
were dropped. However, the Makefile.inc was still processing them.
Remove the generation of the .payload file. This also allows
cbfstool to be updated which generates .elf files on extraction,
but the current recipe didn't provide an architecture so the
build failed.

BUG=chromium:595715
BRANCH=none
TEST=Built reef with updated cbfstool. No errors about unsupported
     architecture.

Change-Id: I5efee74d96ce9e3f46143c110c05665cd734e65e
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/354166
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>